refactor(schema-registry): rename ai to gpt
