+++ libs/schema-registry/src/lib/gpt-conversion-complete/index.ts
