+++ libs/schema-registry/src/lib/gpt-conversion-complete/key.ts
