+++ libs/schema-registry/src/lib/gpt-conversion-complete/value.ts
