+++ libs/schema-registry/src/lib/gpt-conversion-start/index.ts
