+++ libs/schema-registry/src/lib/gpt-conversion-start/key.ts
