+++ libs/schema-registry/src/lib/gpt-conversion-start/value.ts
